requirements.txt
slack-sdk==3.21.3
slackeventsapi==3.0.1
flask==2.3.3
python-dotenv==1.0.0
schedule==1.2.0
apscheduler==3.10.4
requests==2.31.0
certifi==2023.7.22 